requirements.txt
Django==3.1.7
django-composite-field==1.1.0
mysqlclient==2.0.3
Pillow==8.1.1
